rename some variables and stuff
