+++ link.py
@@ -1,5 +1,5 @@
 #// SPDX-License-Identifier: blessing
 #// d a r k s h a p e s
 
-from  modules.nnll_37.src import prepare_tags as indexer
-from  modules.nnll_31.src import count_tensors_and_extract_shape as layer_match
+# from  modules.nnll_37.src import prepare_tags as indexer
+# from  modules.nnll_31.src import count_tensors_and_extract_shape as layer_match
