+++ modules/__init__.py
@@ -1,0 +1,34 @@
+
+"""
+##modules index
+
+#### [nnll 00 - lambda-condensed nested dict traversal](https://github.com/darkshapes/nnll/blob/main/modules/nnll_00/src.py#L8)<br>
+#### [nnll 01 - lambda-condensed nested dict traversal:](https://github.com/darkshapes/nnll/blob/main/modules/nnll_01/src.py#L8)<br>
+#### [nnll 04 - loading safetensors (based on previous performance analysis)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_04/src.py#L5)<br>
+#### [nnll 05 - loading gguf (based on previous performance analysis)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_05/src.py#L2)<br>
+#### [nnll 06 - basic recursive dict crawler](https://github.com/darkshapes/nnll/blob/main/modules/nnll_06/src.py#L14)<br>
+#### [nnll 07 - nn class/type id system](https://github.com/darkshapes/nnll/blob/main/modules/nnll_07/src.py#L2)<br>
+#### [nnll 08 - seed methods](https://github.com/darkshapes/nnll/blob/main/modules/nnll_08/src.py#L2)<br>
+#### [nnll 09 - token encoder type 1](https://github.com/darkshapes/nnll/blob/main/modules/nnll_09/src.py#L12)<br>
+#### [nnll 13 - system capability agent (incomplete)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_13/src.py#L1)<br>
+#### [nnll 16 - scalable, modular gpu device class system (incomplete)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_16/src.py#L6)<br>
+#### [nnll 17 - example device classes for nll_16 (incomplete)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_17/src.py#L4)<br>
+#### [nnll 18 - alternate token encoder for extra prompt length (incomplete)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_18/src.py#L5)<br>
+#### [nnll 22 - diffusers symlink routine(nnll 19, 21)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_22/src.py#L5)<br>
+#### [nnll 23 - dynamic module constructor (nnll 11, 12, 14, 15, 16)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_23/src.py#L5)<br>
+#### [nnll_24 - nested dictionary path return (nnll 02, 03, 06)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_24/src.py#L5)<br>
+#### [nnll_25 - tensor and hash data, regex comparison ](https://github.com/darkshapes/nnll/blob/main/modules/nnll_25/src.py#L9)<br>
+#### [nnll_26 - pytorch-specific seed and tensor routines](https://github.com/darkshapes/nnll/blob/main/modules/nnll_26/src.py)<br>
+#### [nnll_27 - pretty column-formatted console printer](https://github.com/darkshapes/nnll/blob/main/modules/nnll_27/src.py#L6)<br>
+#### [nnll_28 - torch pickletensor loader](https://github.com/darkshapes/nnll/blob/main/modules/nnll_28/src.py#L8)<br>
+#### [nnll_29 - cascading dict comparison filter](https://github.com/darkshapes/nnll/blob/main/modules/nnll_29/src.py#L11)<br>
+#### [nnll 30 - barebones json read/write](https://github.com/darkshapes/nnll/blob/main/modules/nnll_30/src.py#L5)<br>
+#### [nnll 31 - state dict layer name search](https://github.com/darkshapes/nnll/blob/main/modules/nnll_31/src.py#L18)<br>
+#### [nnll 32 - model header extractor (nnll_29)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_32/src.py#L7)<br>
+#### [nnll 33 - layer and tensor value comparison (nnll_24)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_33/src.py#L4)<br>
+#### [nnll 34 - model shard collector (inc)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_34/src.py#L4)<br>
+#### [nnll 35 - filename numeric fetch](https://github.com/darkshapes/nnll/blob/main/modules/nnll_35/src.py#L4)<br>
+#### [nnll 36 - state dict collector](https://github.com/darkshapes/nnll/blob/main/modules/nnll_36/src.py#L13)<br>
+#### [nnll_37 - model indexer(inc)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_37/src.py#L11)<br>
+#### [nnll_38 - tensor_shape)](https://github.com/darkshapes/nnll/blob/main/modules/nnll_38/src.py#L11)<br>
+"""
