+++ modules/nnll_00/src.py
@@ -3,6 +3,7 @@
 
 
 def get_matching_layer(json_file_dict):
+    ## IDE Link
     """
     Test of effectiveness with `map` `lambda` function for nested `dict` path retrieval\n
     :param json_file_dict: `dict` The nested dictionary to parse
