+++ modules/nnll_34/test_src.py
@@ -3,6 +3,25 @@
 
 import unittest
 
+from modules.nnll_34.src import gather_sharded_files
+
+# Note: not an actual test yet
+
+file_paths = [
+    'model_00001-of-00003.safetensors',
+    'model_00002-of-00003.safetensors',
+    'model_00003-of-00003.safetensors',
+    'model_00001-of-00002.gzip',
+    'model_00002-of-00002.gzip',
+    'model_1of3.tar.gz',
+    'model_2of3.tar.gz',
+    'model_3of3.tar.gz'
+]
+results = (gather_sharded_files(p) for p in file_paths)
+for result in results:
+    print(result)
+
+    #raise #dont forget to fill this file
 
 if __name__ == '__main__':
     unittest.main()
