+++ pyproject.toml
@@ -8,11 +8,12 @@ description = "Experimental ML modules library"
 authors = [
     { name = "darkshapes", email = "91800957+exdysa@users.noreply.github.com" },
 ]
+
 requires-python = ">= 3.10"
-license = { file = "LICENSE" }
-readme = "README.md"
-urls = { source = "https://github.com/darkshapes/nnll" }
-dynamic = ["version"]
+license         = { file = "LICENSE" }
+readme          = "README.md"
+urls            = { source = "https://github.com/darkshapes/nnll" }
+dynamic         = ["version"]
 
 [project.optional-dependencies]
 opt = [
@@ -23,16 +24,8 @@ opt = [
     "diffusers",
     "transformers",
 ]
-dev = [
-    "pytest",
-    "huggingface_hub",
-    "numpy",
-    "torch",
-    "llama-cpp-python",
-    "sentencepiece",
-    "diffusers",
-    "transformers",
-]
+
+dev = ["nnll[opt]", "pytest", "huggingface_hub"]
 
 [project.scripts]
 semver = "main:main"
@@ -44,4 +37,4 @@ dependencies = { file = ["requirements.txt"] }
 write_to = "_version.py"
 
 [tool.uv]
-dev-dependencies = ["pytest"]
+dev-dependencies = ["nnll[opt]", "nnll[dev]"]
